+++ nerd/nerd_client.py
@@ -7,7 +7,7 @@ except ImportError:
 
 from zenlog import logging
 
-from .client import ApiClient
+from nerd.client import ApiClient
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
